pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hydradb-bench"
version = "0.1.0"
description = "Complete benchmark framework for HydraDB using RAGAS"
requires-python = ">=3.10"
dependencies = [
    "ragas>=0.2.0",
    "langchain-openai>=0.2.0",
    "langchain>=0.3.0",
    "httpx>=0.27.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "jinja2>=3.1",
    "pandas>=2.0",
    "rich>=13.0",
    "nest-asyncio>=1.6.0",
    "datasets>=2.14.0",
    "tiktoken>=0.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.23",
    "pytest-mock>=3.12",
]

[project.scripts]
hydradb-bench = "hydradb_bench.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
